+++ FTAnalyser/Core/Displays/IDisplayColourBMD.cs
@@ -29,6 +29,6 @@ namespace FTAnalyzer
         FactLocation BirthLocation { get; }
         FactLocation DeathLocation { get; }
         FactLocation BestLocation(FactDate when);
-        int Ahnentafel { get; }
+        Int64 Ahnentafel { get; }
     }
 }
